Manage Accounts Almost Finished

+++ PayMonitorUI/FrmAccounts.cs
@@ -24,51 +24,98 @@ namespace PayMonitorUI
         SqlCommand cmd;
         SqlConnection conn;
         SqlCommand checkercmd;
+        SqlDataAdapter adapt;
+        DataTable dt;
+
 
         string connstr = ConfigurationManager.ConnectionStrings["PayMonitorDB"].ConnectionString;
 
         private void FrmAccounts_Load(object sender, EventArgs e)
         {
+            //conn = new SqlConnection(connstr);
+            //conn.Open();
+            //adapt = new SqlDataAdapter("select * from tbl_accounts", conn);
+            //dt = new DataTable();
+            //adapt.Fill(dt);
+            //grdViewAccounts.DataSource = dt;
+            //conn.Close();
+            Load_ViewRegisteredData();
+            AutoIncrementIDForm();
+            
+            
+
+        }
+
+        public void AutoIncrementIDForm()
+        {
+            // Force Auto Increment Value
+            SqlDataReader datareader;
+            try
+            {
+                conn.Open();
+                checkercmd = new SqlCommand("SELECT top 1 * FROM tbl_accounts order by account_id Desc;", conn);
+                datareader = checkercmd.ExecuteReader();
+
+                if (datareader.HasRows)
+                {
+                    while (datareader.Read())
+                    {
+                        string lastIDString = datareader["account_id"].ToString();
+                        int lastId = Int32.Parse(lastIDString);
+                        lastId += 1;
+                        txtAccountID.Text = lastId.ToString();
+                        txtAccountID.Enabled = false;
+                    }
+                    datareader.Close();
+                }
+                conn.Close();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
 
         }
 
         private void btnRegister_Click(object sender, EventArgs e)
         {
-                 conn = new SqlConnection(connstr);
-                 cmd = new SqlCommand("INSERT INTO tbl_accounts (account_id, role_name, username, password, firstname, lastname) VALUES (@accountID, @rolename, @username, @password, @firstname, @lastname)", conn);
+            conn = new SqlConnection(connstr);
+            //cmd.CommandText = "SET IDENTITY_INSERT tbl_accounts ON";
+            cmd = new SqlCommand("INSERT INTO tbl_accounts (account_id, role_name, username, password, firstname, lastname) VALUES (@accountID, @rolename, @username, @password, @firstname, @lastname)", conn);
 
-                 cmd.Parameters.AddWithValue("@accountID", txtAccountID.Text);
-                 cmd.Parameters.AddWithValue("@rolename", cmbAccountType.SelectedItem);
-                 cmd.Parameters.AddWithValue("@username", txtUsername.Text);
-                 cmd.Parameters.AddWithValue("@password", txtPassword.Text);
-                 cmd.Parameters.AddWithValue("@firstname", txtFirstname.Text);
-                 cmd.Parameters.AddWithValue("@lastname", txtLastname.Text);
+            cmd.Parameters.AddWithValue("@accountID", txtAccountID.Text);
+            cmd.Parameters.AddWithValue("@rolename", cmbAccountType.SelectedItem);
+            cmd.Parameters.AddWithValue("@username", txtUsername.Text);
+            cmd.Parameters.AddWithValue("@password", txtPassword.Text);
+            cmd.Parameters.AddWithValue("@firstname", txtFirstname.Text);
+            cmd.Parameters.AddWithValue("@lastname", txtLastname.Text);
 
-                 checkercmd = new SqlCommand("SELECT * FROM tbl_accounts WHERE account_id=@accountID OR username=@username", conn);
-                 checkercmd.Parameters.AddWithValue("@accountID", txtAccountID.Text);
-                 checkercmd.Parameters.AddWithValue("@username", txtUsername.Text);
+            checkercmd = new SqlCommand("SELECT * FROM tbl_accounts WHERE account_id=@accountID OR username=@username", conn);
+            checkercmd.Parameters.AddWithValue("@accountID", txtAccountID.Text);
+            checkercmd.Parameters.AddWithValue("@username", txtUsername.Text);
 
-                 try
-                 {
-                     SqlDataReader datareader;
+            try
+            {
+                SqlDataReader datareader;
 
-                     conn.Open();
-                     datareader = checkercmd.ExecuteReader();
-                     if (datareader.HasRows)
-                     {
-                         MessageBox.Show("The " + txtUsername.Text + " username is taken. It has user id =" + txtAccountID.Text + ".");
-                         datareader.Close();
-                     }
-                     else
-                     {
-                         datareader.Close();
-                         if (cmd.ExecuteNonQuery() == 1)
-                         {
-                             MessageBox.Show("User Successfully Added!");
-                         }
-                     }
-                     Load_ViewRegisteredData();
-                     conn.Close();
+                conn.Open();
+                datareader = checkercmd.ExecuteReader();
+
+                if (datareader.HasRows)
+                {
+                    MessageBox.Show("The " + txtUsername.Text + " username is taken. It has user id =" + txtAccountID.Text + ".");
+                    datareader.Close();
+                }
+                else
+                {
+                    datareader.Close();
+                    if (cmd.ExecuteNonQuery() == 1)
+                    {
+                        MessageBox.Show("User Successfully Added!");
+                    }
+                }
+                Load_ViewRegisteredData();
+                conn.Close();
             }
 
             catch (Exception ex)
@@ -122,23 +169,31 @@ namespace PayMonitorUI
         private void btnDelete_Click(object sender, EventArgs e)
         {
             conn = new SqlConnection(connstr);
-            cmd = new SqlCommand("DELETE FROM tbl_accounts WHERE username=@username", conn);
+            cmd = new SqlCommand("DELETE FROM tbl_accounts WHERE account_id=@account_id OR username=@username", conn);
+            cmd.Parameters.AddWithValue("@account_id", txtAccountID.Text);
             cmd.Parameters.AddWithValue("@username", txtUsername.Text);
 
             try
             {
                 conn.Open();
-                if (cmd.ExecuteNonQuery() == 1)
+                if (txtAccountID.Text == "1" || txtAccountID.Text == "0")
                 {
-                    MessageBox.Show("Data Deleted Sucessfully!");
-                    Load_ViewRegisteredData();
-                    txtAccountID.Text = "";
-                    cmbAccountType.Text = "";
-                    txtFirstname.Text = "";
-                    txtLastname.Text = "";
-                    txtPassword.Text = "";
-                    txtUsername.Text = "";
-
+                    // Don't Allow to delete super administrator.
+                    MessageBox.Show("You don't have a permission to delete this administrator");
+                }
+                else
+                {
+                    if (cmd.ExecuteNonQuery() == 1)
+                    {
+                        MessageBox.Show("Data Deleted Sucessfully!");
+                        Load_ViewRegisteredData();
+                        txtAccountID.Text = "";
+                        cmbAccountType.Text = "";
+                        txtFirstname.Text = "";
+                        txtLastname.Text = "";
+                        txtPassword.Text = "";
+                        txtUsername.Text = "";
+                    }
                 }
 
                 conn.Close();
@@ -175,11 +230,13 @@ namespace PayMonitorUI
             txtFirstname.Text = "";
             txtLastname.Text = "";
 
+            // Force Auto Increment Value
+            AutoIncrementIDForm();
         }
 
         private void btnUpdate_Click(object sender, EventArgs e)
         {
-            conn = new SqlConnection(Properties.Settings.Default.PayMonitorDB);
+            conn = new SqlConnection(connstr);
             cmd = new SqlCommand("UPDATE tbl_accounts SET account_id=@accountID, role_name=@rolename, username=@username, password=@password, firstname=@firstname, lastname=@lastname WHERE account_id=@accountID OR username=@username", conn);
 
             cmd.Parameters.AddWithValue("@accountID", txtAccountID.Text);
@@ -206,7 +263,22 @@ namespace PayMonitorUI
             }
         }
 
+        private void txtSearch_TextChanged(object sender, EventArgs e)
+        {
+            {
+                conn = new SqlConnection(connstr);
+                conn.Open();
+                // search anything by account_id, username, firstname, last_name, role_name
+                string qry = $"SELECT * FROM tbl_accounts WHERE account_id LIKE '%{txtSearch.Text}%' OR username LIKE '%{txtSearch.Text}%' OR firstname LIKE '%{txtSearch.Text}%' OR lastname LIKE '%{txtSearch.Text}%' OR role_name LIKE '%{txtSearch.Text}%'";
+                adapt = new SqlDataAdapter(qry, conn);
+                dt = new DataTable();
+                adapt.Fill(dt);
+                grdViewAccounts.DataSource = dt;
+                conn.Close();
+            }
+        }
     }
+
 }
